+++ processor/src/test/fixtures/expected/com/example/fragment/types/Sting_TypedProviderMethodModel.java
@@ -1,0 +1,10 @@
+package com.example.fragment.types;
+
+import javax.annotation.processing.Generated;
+
+@Generated("sting.processor.StingProcessor")
+public final class Sting_TypedProviderMethodModel implements TypedProviderMethodModel {
+  public MyModel $sting$_provideMyModel() {
+    return provideMyModel();
+  }
+}