+++ processor/src/test/fixtures/input/com/example/fragment/types/TypedProviderMethodModel.java
@@ -1,0 +1,14 @@
+package com.example.fragment.types;
+
+import sting.Fragment;
+import sting.Typed;
+
+@Fragment
+public interface TypedProviderMethodModel
+{
+  @Typed( Runnable.class )
+  default MyModel provideMyModel()
+  {
+    return null;
+  }
+}