+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Explicitly document the expectations for `@StingProvider` types that contain `@Eager`, `@Named` and `@Typed` annotations.
+* Explicitly document and add integration tests to ensure that `@Typed` annotations can appear on provider methods of fragments.
 
 ### [v0.32](https://github.com/sting-ioc/sting/tree/v0.32) (2026-04-29) · [Full Changelog](https://github.com/sting-ioc/sting/compare/v0.31...v0.32)